+++ src/main/java/kg/peaksoft/ebookm1/entity/book/AudioBook.java
@@ -1,4 +1,4 @@
-package kg.peaksoft.ebookm1.entity;
+package kg.peaksoft.ebookm1.entity.book;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
