+++ src/main/java/kg/peaksoft/ebookm1/entity/book/FavoritesBooks.java
@@ -1,31 +1,26 @@
-package kg.peaksoft.ebookm1.entity;
+package kg.peaksoft.ebookm1.entity.book;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import org.springframework.format.annotation.DateTimeFormat;
 
 import javax.persistence.*;
-import java.time.LocalDate;
 import java.util.List;
+
 @Entity
-@Table(name = "eBooks")
+@Table(name = "favoritesBooks")
 @Setter
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-public class eBook {
+public class FavoritesBooks {
 
     @Id
     @GeneratedValue(strategy = GenerationType.SEQUENCE)
-    @Column(name = "eBook_id")
     private Long id;
-    private String bookFragment;
-    @Transient
-    private eBook eBook;
+    @OneToMany
+    private List<Book> books;
 
 
-
-    }
-
+}
