+++ src/main/java/kg/peaksoft/ebookm1/entity/enumClass/Language.java
@@ -1,4 +1,4 @@
-package kg.peaksoft.ebookm1.entity;
+package kg.peaksoft.ebookm1.entity.enumClass;
 
 public enum Language {
     ENGLISH,
