+++ src/main/java/kg/peaksoft/ebookm1/entity/otherClass/Address.java
@@ -1,4 +1,5 @@
-package kg.peaksoft.ebookm1.entity;
+package kg.peaksoft.ebookm1.entity.otherClass;
+
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -6,21 +7,20 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import javax.persistence.*;
-import java.util.List;
 
 @Entity
-@Table(name = "favoritesBooks")
+@Table(name = "addresses")
 @Setter
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-public class FavoritesBooks {
+public class Address {
 
     @Id
     @GeneratedValue(strategy = GenerationType.SEQUENCE)
     private Long id;
-    @OneToMany
-    private List<Book> books;
-
-
+    private String country;
+    private String city;
+    private String street;
+    private int index;
 }
