+++ src/main/java/kg/peaksoft/ebookm1/entity/otherClass/Search.java
@@ -1,28 +1,26 @@
-package kg.peaksoft.ebookm1.entity;
+package kg.peaksoft.ebookm1.entity.otherClass;
 
+import kg.peaksoft.ebookm1.entity.book.Book;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import javax.persistence.*;
-import java.time.LocalDateTime;
 import java.util.List;
 
 @Entity
-@Table(name = "promoCodes")
+@Table(name = "searches")
 @Setter
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-public class PromoCode {
+public class Search {
     @Id
     @GeneratedValue(strategy = GenerationType.SEQUENCE)
     private Long id;
-    private String promoName;
-    private LocalDateTime startingDay;
-    private LocalDateTime finishingDay;
-    private byte percent;
+    private String search;
     @OneToMany
     private List<Book> books;
+
 }
